flask/requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
Flask
python-dotenv
tiktoken
langchain-openai